# Remove look-up optimizations in `CountBy`/`ScanBy`