@atifaziz 2022-10-26
> they were not an actual performance improvement.

The shame is that [at the time this was benchmarked](https://github.com/morelinq/MoreLINQ/pull/207#issuecomment-258132466), only the optimised version was tested against a sorted and unsorted source sequence, but no results were published against an unoptimised version (I imagine but can't remember if I did that locally but just didn't publish). I did not go and test against versions prior to .NET Framework 4.8 so I cannot say if some other things were improved (in the dictionary) since then that make optimisation redundant. I reckon some of the loss is coming from hashing twice when the key changes, which could be often for an unsorted source, but that's just a guess.